requirements.txt
langchain
openai==0.28
langchain_community
huggingface_hub
python-dotenv
streamlit